Change to use Dependecy injection and Services
