+++ Backend/RealManager/RealManager/Controllers/MatchController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using RealManager.Domain;
+using RealManager.Services.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -11,83 +12,17 @@ namespace RealManager.Controllers
     [Route("[controller]")]
     public class MatchController : ControllerBase
     {
-        public MatchController()
+        private readonly IMatchService _matchService;
+        public MatchController(IMatchService matchService)
         {
-            
+            _matchService = matchService;
         }
 
         [HttpGet]
         [Route("/{time}")]
         public bool Get([FromRoute]int time)
         {
-            var attackPlayer = new Player()
-            {
-                Defence = 80,
-                Drible = 80,
-                Pace = 80,
-                Pass = 80,
-                Physical = 80,
-                Shoot = 80
-            };
-
-            var defencePlayer = new Player()
-            {
-                Defence = 80,
-                Drible = 80,
-                Pace = 80,
-                Pass = 80,
-                Physical = 20,
-                Shoot = 80
-            };
-
-            var goalKeeperPlayer = new Player()
-            {
-                Defence = 80,
-                Drible = 80,
-                Pace = 80,
-                Pass = 80,
-                Physical = 20,
-                Shoot = 80
-            };
-
-            int dribleSuccessChance = attackPlayer.Drible + attackPlayer.Pace / 2 - 15 * time / attackPlayer.Physical;
-            var randomNumber = new Random();
-
-            var realDrible = randomNumber.Next(0, 150);
-
-            var ableToDrible = realDrible > 150 - dribleSuccessChance;
-
-            if (!ableToDrible) return false;
-
-
-            int tackleSucessChance = defencePlayer.Defence + defencePlayer.Pace / 2 - 15 * time / defencePlayer.Physical;
-
-            var realTackle = randomNumber.Next(0, 150);
-
-            var ableToTackle = realTackle > 150 - tackleSucessChance;
-
-            if (ableToTackle) return false;
-
-
-            int shootOnTargetChance = attackPlayer.Shoot - 15 * time / attackPlayer.Physical;
-
-            var realShoot = randomNumber.Next(0, 100);
-
-            var ableToShootOnTarget = realShoot > 100 - shootOnTargetChance;
-
-            if (!ableToShootOnTarget) return false;
-
-
-
-            int defenceSucessChance = goalKeeperPlayer.Defence + goalKeeperPlayer.Pace / 2 - 5 * time / goalKeeperPlayer.Physical;
-
-            var realDefence = randomNumber.Next(0, 150);
-
-            var ableToDefence = realDefence > 150 - defenceSucessChance;
-
-            if (ableToDefence) return false;
-
-            return true;
+            return _matchService.RunMatchEvent(time);
         }
     }
 }
